Change 1-of-3 to 2-of-3 multisig
One of the two previous proposals was a 1-of-3 multisig where the
third public key would be artificially generated along with proof
that a corresponding private key is unknown and would play the role
of the TIR transaction tag. This however would be susceptible to a
lucky attacker who could find a corresponding private key and be
able to spend all TIR transactions.

Christos proposed to replace the above with a 2-of-3 multisig
with a publicly known private key for the tag. This way TIR
transactions are future-proof against the previously described
atttack and correctly emulate the 1-of-2 desired behavior.