Add network data to generalize pubKeyToEntity()
Using Address.fromHash() without a network argument returns a main
net address, which is not always desired. Thus pubKeyToEntity() now
takes the type of network as an argument and passes it to
Address.fromHash(). All cases where this function is used are
expanded with the new input. Additionally, the DirectTrust class is
expanded with the network property for use in pubKeyToEntity()
calls.