Let spv node decrease trust correctly
Use logic similar to createTrustIncreasingMTX() for
createTrustDecreasingMTX(s)(). This way only an spv node passes the
wallet and thus the former function can correctly decide whether to use
the wallet to obtain the coin. Using the wallet is necessary in the case
of an spv node, because this kind of node does not have the getCoin()
method.